+++ app/scripts/lib/snap-keyring/snap-keyring.ts
@@ -19,10 +19,8 @@ import { t } from '../../translate';
 import { IconName } from '../../../../ui/components/component-library/icon';
 import MetaMetricsController from '../../controllers/metametrics-controller';
 import { getUniqueAccountName } from '../../../../shared/lib/accounts';
-import {
-  getSnapName,
-  isSnapPreinstalled,
-} from '../../../../shared/lib/snaps/snaps';
+import { isSnapPreinstalled } from '../../../../shared/lib/snaps/snaps';
+import { getSnapName } from '../../../../shared/lib/accounts/snaps';
 import { SnapKeyringBuilderMessenger } from './types';
 import { isBlockedUrl } from './utils/isBlockedUrl';
 import { showError, showSuccess } from './utils/showResult';
